+++ CHANGELOG.md
@@ -10,6 +10,10 @@
 
 - Added methods [`FiniteBifield.multiply`](https://komm.dev/ref/FiniteBifield/#multiply), [`FiniteBifield.divide`](https://komm.dev/ref/FiniteBifield/#divide), and [`FiniteBifield.power`](https://komm.dev/ref/FiniteBifield/#power), which operate elementwise on arrays of integer representations of field elements.
 
+### Breaking changes
+
+- Removed method `bch_syndrome` from `BCHCode`. Instead of `code.bch_syndrome(r_poly)`, use `[r_poly.evaluate(code.alpha**i) for i in range(1, code.delta)]`.
+
 ### Fixed
 
 - Changed `BerlekampDecoder` to return the received word when the error locator has too few roots.